extract out generateQuestions to helpers
